A Fucked Sieve of Eratosthenes
The prof has asked me to write a Sieve of Eratosthenes without using any of:
* Arrays
* Recursion
* Strings
* I/O

for an arbitrarily large N.

As far as I can tell, this is impossible. I have instead written a Sieve-*like* algorithm that will work and find
primes for all values of N < 100 (and have set N = 100 by default) but as directly written the problem is
impossible to solve.

Signed-off-by: Connor McDermid <mcdermidconnor@outlook.com>